replace m_diagonal by DiagonalCCW defined in seims.h

+++ seims/src/seims_main/modules/hydrology/CH_DW/DiffusiveWave.cpp
@@ -204,7 +204,7 @@ void DiffusiveWave::initialOutputs() {
                 dx = m_CellWidth / cos(atan(s0));
                 int dir = (int) m_direction[id];
                 //if ((int) m_diagonal[dir] == 1) {
-                if ((int) DiagonalCCW[dir] == 1) {
+                if (DiagonalCCW[dir] == 1) {
                     dx = SQ2 * dx;
                 }
                 m_flowLen[i][j] = dx;
